pyproject.toml
[tool.poetry]
name = "The Ranny Parkour"
version = "0.1.0"
description = ""
license = "MIT"
authors = ["lew1101 <kennethlew001@gmail.com>", "Junlin Song <jsong@wearelcc.ca>", "Aryan Tavassoli <atavassoli@wearelcc.ca>"]

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.1.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
cx-Freeze = "^6.10"
autopep8 = "^1.6.0"
